:sparkles: ABI v17 in the staticlib: HostOps::guest_ram_regions, and the two declarations of a span pinned against each other
Component: `crates/reims-vgpu` — `qemu::abi`, `qemu::host_ops`, `runtime::host`,
`runtime::guest_ram`, and the shared ABI header. Both pathways; no behavior
change yet, because nothing calls the new entry point.

Bumps `REIMS_VGPU_QEMU_ABI_VERSION` 16 -> 17 and adds the v17 history entry to
`include/reims_vgpu_qemu_abi.h` in the established style — what it adds, and
why the shim exports the answer rather than the inputs.

What crosses:

- `ReimsVgpuGuestRamRegion` / `runtime::guest_ram::GuestRamRegion`, three
  `uint64_t`s. `host_va` is a `u64` rather than a `usize` so the layout does
  not depend on the target the shim happened to be built for.
- `REIMS_VGPU_GUEST_RAM_ERR_ARGS` / `_NO_RAM`.
- `ReimsVgpuHostOps.guest_ram_regions`, placed after `dmabuf_for_pages` in
  both declarations.

Every one of those exists twice with nothing in the toolchain comparing them,
so both new `header_define` tests are here.
`the_abi_header_agrees_on_the_guest_ram_region_layout` checks the C field
names and their order, not just the size: a reordered field is not a decode
error, it is a host address read as a length and imported as a span, and the
bound in `runtime::guest_ram` cannot catch that because every number it checks
would be self-consistent.

The Rust adapter grows its array from the shim's own reported total rather
than capping at its first guess. A device that imported the first eight spans
and dropped the rest would run the copying rails for part of the guest's RAM
with nothing in the log saying which part — the "no silent caps" rule, at the
one call that decides how much of guest memory the GPU can reach at all. A
total that never converges is a named refusal, not a retry loop.

Tests: 5 new adapter tests (missing callback refuses by name; a fitting
machine is asked exactly once; 21 spans against a first array of 8 still
yields all 21 in two calls; a never-converging total refuses with both
numbers; each negative code keeps its own check including an unknown one) and
2 new `header_define` tests. `cargo test -p reims-vgpu --no-default-features
--features backend-vulkan,host-window -- --test-threads=1`: 1647 lib + all
integration green, ignored counts unchanged.

Clippy `-D warnings` clean on the native Vulkan arm and on `--target
aarch64-apple-darwin --features backend-metal`.

Not verified: no live boot on either pathway; the `backend-metal` test arm was
not run (Linux host).

+++ crates/reims-vgpu/src/runtime/guest_ram.rs
@@ -78,12 +78,21 @@ use crate::observe::{Decline, Emit};
 /// `map_pages` answers "give me a view of these specific pages" and may build a
 /// transient one the caller must release; this answers "where does guest RAM
 /// live, for the lifetime of the VM".
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+///
+/// `#[repr(C)]` and three `u64`s because the shim writes these directly through
+/// the caller's array: this declaration and `ReimsVgpuGuestRamRegion` in
+/// `include/reims_vgpu_qemu_abi.h` are one struct, and
+/// `crate::qemu::abi::tests::the_abi_header_agrees_on_the_guest_ram_region_layout`
+/// is the only thing that compares them. The host address is a `u64` rather
+/// than a `usize` so the layout does not depend on the target the shim happened
+/// to be built for.
+#[repr(C)]
+#[derive(Clone, Copy, Debug, Default, PartialEq, Eq)]
 pub struct GuestRamRegion {
     /// First guest physical address this block backs.
     pub gpa_base: u64,
     /// Host virtual address QEMU mapped it at. Stable for the VM's lifetime.
-    pub host_va: usize,
+    pub host_va: u64,
     /// Length in bytes, in both address spaces.
     pub len: u64,
 }
@@ -145,7 +154,7 @@ pub enum GuestRamError {
     /// `host_va + len` or `gpa_base + len` leaves its address space. A RAMBlock
     /// cannot wrap, so this is a malformed answer from the shim rather than a
     /// large one.
-    RegionWraps { host_va: usize, len: u64 },
+    RegionWraps { host_va: u64, len: u64 },
     /// The backend reported an import granularity that is zero or not a power of
     /// two. Every alignment computation below assumes a power of two mask, so
     /// this is refused rather than worked around.
@@ -313,7 +322,8 @@ impl GuestRamImport {
             host_va: region.host_va,
             len: region.len,
         };
-        let host_end = (region.host_va as u64)
+        let host_end = region
+            .host_va
             .checked_add(region.len)
             .ok_or(wraps)
             .map_err(GuestRamError::report)?;
@@ -330,10 +340,10 @@ impl GuestRamImport {
             align,
             len: region.len,
         };
-        let host_base = align_up_u64(region.host_va as u64, align)
+        let host_base = align_up_u64(region.host_va, align)
             .ok_or(unsatisfiable)
             .map_err(GuestRamError::report)?;
-        let head = host_base - region.host_va as u64;
+        let head = host_base - region.host_va;
         let len = region
             .len
             .checked_sub(head)
@@ -781,7 +791,7 @@ mod tests {
         assert!(matches!(
             GuestRamImport::new(
                 GuestRamRegion {
-                    host_va: usize::MAX - 0xfff,
+                    host_va: u64::MAX - 0xfff,
                     len: 0x4000,
                     ..ok
                 },
